+++ pkg/utils/values.go
@@ -2,7 +2,7 @@ package utils
 
 import "strings"
 
-const VERSION = "0.0.1"
+const VERSION = "0.0.2"
 
 var PROVIDER = ""
 var GITHUBNAME = ""