+++ CHANGELOG.md
@@ -2,7 +2,20 @@
 
 ---
 
+## [v0.0.2](https://github.com/phil1436/GitGitGo-CLI/tree/0.0.2) (2023-11-26)
+
+_PRE-RELEASE_
+
+-   Bug fixes
+-   Description and Keywords support for `print` command
+-   `.gitgitgoc` file support
+-   keyword support for `init` command
+
+---
+
 ## [v0.0.1](https://github.com/phil1436/GitGitGo-CLI/tree/0.0.1) (2023-11-8)
+
+_PRE-RELEASE_
 
 -   _Initial release_